Skeleton refactor: Put drawXGrid and drawYGrid in the Skeleton class
